+++ src/main/java/net/oneandone/pommes/repository/GiteaRepository.java
@@ -21,23 +21,18 @@ import io.gitea.Configuration;
 import io.gitea.api.OrganizationApi;
 import io.gitea.api.RepositoryApi;
 import io.gitea.auth.ApiKeyAuth;
-import io.gitea.model.ContentsResponse;
 import io.gitea.model.Organization;
 import net.oneandone.inline.Console;
 import net.oneandone.pommes.cli.Environment;
 import net.oneandone.pommes.descriptor.Descriptor;
-import net.oneandone.pommes.scm.GitUrl;
-import net.oneandone.sushi.fs.NodeInstantiationException;
 import net.oneandone.sushi.fs.World;
-import net.oneandone.sushi.fs.memory.MemoryNode;
 import net.oneandone.sushi.util.Strings;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
@@ -47,15 +42,13 @@ import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 
 /** https://developer.atlassian.com/static/rest/bitbucket-server/4.6.2/bitbucket-rest.html */
-public class GiteaRepository extends Repository {
-    private static final String PROTOCOL = "gitea:";
-
+public class GiteaRepository extends Repository<GiteaProject> {
     public static void main(String[] args) throws IOException, URISyntaxException, InterruptedException {
         Environment env;
         GiteaRepository gitea;
 
         env = new Environment(Console.create(), World.create());
-        gitea = GiteaRepository.create(env, "reponame", "https://git.ionos.org/CPOPS");
+        gitea = GiteaRepository.create(env, "reponame", "https://git.ionos.org/CPOPS", null);
         BlockingQueue<Descriptor> result = new ArrayBlockingQueue<>(1000);
         gitea.scan(result, env.console());
         for (var d : result) {
@@ -63,15 +56,8 @@ public class GiteaRepository extends Repository {
         }
     }
 
-    public static GiteaRepository createOpt(Environment environment, String repository, String url) throws URISyntaxException, NodeInstantiationException {
-        if (url.startsWith(PROTOCOL)) {
-            return create(environment, repository, url.substring(PROTOCOL.length()));
-        } else {
-            return null;
-        }
-    }
-
-    public static GiteaRepository create(Environment environment, String repository, String uriStr) throws URISyntaxException {
+    public static GiteaRepository create(Environment environment, String repository, String uriStr, PrintWriter log)
+            throws URISyntaxException {
         ApiClient gitea;
         URI uri;
         String path;
@@ -116,9 +102,8 @@ public class GiteaRepository extends Repository {
     }
 
     @Override
-    public void scan(BlockingQueue<Descriptor> dest, Console console) throws IOException, InterruptedException {
+    public List<GiteaProject> list() throws IOException {
         List<String> orgs;
-        Descriptor descriptor;
         orgs = listCurrentUserOrgs();
         orgs.addAll(listOrganizations());
 
@@ -129,42 +114,22 @@ public class GiteaRepository extends Repository {
         } else {
             orgs = Collections.singletonList(selectedOrganization);
         }
+        List<GiteaProject> result = new ArrayList<>();
         for (String org : orgs) {
             for (var r : listRepos(org)) {
-                descriptor = scanOrganizationOpt(org, r.getName(), r.getDefaultBranch());
-                if (descriptor != null) {
-                    dest.put(descriptor);
-                }
+                result.add(new GiteaProject(org, r.getName(), r.getDefaultBranch()));
             }
         }
+        return result;
     }
 
-    private Descriptor scanOrganizationOpt(String org, String repo, String ref) throws IOException {
-        List<ContentsResponse> lst;
-
+    @Override
+    public Descriptor load(GiteaProject project) throws IOException {
         try {
-            lst = repositoryApi.repoGetContentsList(org, repo, ref);
-            for (ContentsResponse contents : lst) {
-                if ("file".equals(contents.getType())) {
-                    Descriptor.Creator m = Descriptor.match(contents.getName());
-                    if (m != null) {
-                        contents = repositoryApi.repoGetContents(org, repo, contents.getPath(), ref);
-                        String str = contents.getContent();
-                        if ("base64".equals(contents.getEncoding())) {
-                            str = new String(Base64.getDecoder().decode(str.getBytes(StandardCharsets.UTF_8)), StandardCharsets.UTF_8);
-                        } else if (contents.getEncoding() != null) {
-                            throw new IllegalStateException(contents.getEncoding());
-                        }
-                        MemoryNode tmp = environment.world().memoryNode(str);
-                        return m.create(environment, tmp, repo, org + "/" + repo + "/" + contents.getPath(), tmp.sha(),
-                                GitUrl.create("ssh://gitea@" + hostname + "/" + org + "/" + repo + ".git"));
-                    }
-                }
-            }
+            return project.scanOpt(environment, hostname, repositoryApi);
         } catch (ApiException e) {
             throw new IOException(e);
         }
-        return null;
     }
 
     public List<io.gitea.model.Repository> listRepos(String org) throws IOException {
@@ -231,7 +196,4 @@ public class GiteaRepository extends Repository {
         }
         return result;
     }
-
-
-
 }